Moved video player code into own file. Improved video type and element detection. Fixed overlapping controls for Stories. Unfinished: Broke auto-un-mute and volume storage.
