+++ content/style.css
@@ -17,6 +17,10 @@
     background: rgba(0, 0, 0, 0.35);
 }
 
+.videoControlsInStory {
+    background: rgba(38,38,38,.6);
+}
+
 .videoControlElement {
     margin: 4px;
 }
