+++ src/videoType.ts
@@ -1,0 +1,13 @@
+export enum VideoType {
+    // The video is on the default main feed.
+    feed,
+
+    // The video is in a Reel.
+    reel,
+
+    // The video is in a Story.
+    story,
+
+    // The video is on the Explore grid.
+    explore
+}
